cleanup(cvvdp-gpu): clear remaining clippy warnings on tests + examples
Tick 232: after tick 231 closed the library lints, --all-targets
still flagged 7 in tests/examples.

Fixes:
- tests/common/mod.rs: collapsed nested if-let into a let-chain
  using `&&` (stable since MSRV 1.85).
- tests/pipeline_color.rs:
    - Dropped redundant `(wu * hu) as usize` cast (wu/hu already usize)
    - Module-level `#![allow(clippy::needless_range_loop)]` for the
      3 per-band loops where k indexes ref_tp[k] / d_bands[k] plus
      side metadata — converting to enumerate is a wash. Mirrors
      the library's same allow.
- tests/cpu_backend.rs: `#![allow(clippy::excessive_precision)]` for
  the 7-digit pycvvdp golden literal. Same rationale as the lib-
  level allow — the literal documents the source value even though
  LLVM rounds at f32.
- examples/manifest_parity_probe.rs: same allow attribute for the
  6 pycvvdp golden literals in the fixture table.

Verified:
- cargo clippy -p cvvdp-gpu --features cuda --all-targets: clean
- cargo test --test pipeline_color: 27/27 ok
- cargo test --test pipeline_score: 9/9 ok
- cargo test --test cpu_backend: 4/4 ok

CHANGELOG.md `[Unreleased]` → cvvdp-gpu (cleanup) entry added.

+++ crates/cvvdp-gpu/tests/pipeline_color.rs
@@ -3,6 +3,11 @@
 //! pipeline. Compares against the host scalar `srgb_byte_to_dkl_scalar`.
 
 #![cfg(any(feature = "cuda", feature = "wgpu", feature = "hip"))]
+// Per-band loops naturally use `for k in 0..n_bands` indexing into
+// `ref_tp[k]` / `d_bands[k]` plus side metadata (sentinels,
+// per-band widths) — converting to enumerate is a wash. Same
+// pattern as the library's `#![allow(clippy::needless_range_loop)]`.
+#![allow(clippy::needless_range_loop)]
 
 use cubecl::Runtime;
 use cvvdp_gpu::Cvvdp;
@@ -1540,7 +1545,7 @@ fn compute_dkl_t_p_bands_matches_host_scalar_per_pixel_at_chroma_shift() {
     // host scalar. Uses the SAME Weber + log_l_bkg the GPU path
     // produces internally (read back via the host scalar Weber
     // pyramid; tick 198 confirmed they match the GPU to 2.7e-7).
-    let n_px = (wu * hu) as usize;
+    let n_px = wu * hu;
     let display = DisplayModel::STANDARD_4K;
     let mut planes: [Vec<f32>; 3] = [vec![0.0; n_px], vec![0.0; n_px], vec![0.0; n_px]];
     for (i, chunk) in ref_srgb.chunks_exact(3).enumerate() {
